# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ScreenPulse API",
    "version": "1.0.0",
    "description": "ScreenPulse API documentation.",
    "contact": {
      "name": "Eduardo González",
      "email": "your@email.com",
      "url": "https://github.com/EduGese/ScreenPulse-backend-Api"
    }
  },
  "servers": [
    {
      "url": "http://localhost:9000",
      "description": "Local server"
    },
    {
      "url": "https://screenpulse-api.onrender.com",
      "description": "Live server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for writing operations on ScreenPulse API. You can obtain it contacting the administrator. Key will be validated when performing requests to endpoints that modify data. For read-only operations, the API key is not required."
      }
    },
...
    "schemas": {
...
      "ValidationError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Validation failed"
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "status": {
            "type": "integer",
            "example": 400
          },
          "errors": {
            "type": "array",
            "description": "Array of validation errors. Each object describes a single validation failure.",
            "items": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Email is invalid"
                },
                "param": {
                  "type": "string",
                  "example": "email"
                },
                "location": {
                  "type": "string",
                  "example": "body"
                },
                "value": {
                  "type": "string",
                  "example": "not-an-email",
                  "nullable": true
                }
              },
              "required": [
                "msg",
                "param",
                "location"
              ]
            }
          }
        },
        "required": [
          "error",
          "code",
          "status",
          "errors"
        ]
      },
      "InternalServerError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Internal server error"
          },
          "code": {
            "type": "string",
            "example": "INTERNAL_ERROR"
          },
          "status": {
            "type": "integer",
            "example": 500
          }
        },
        "required": [
          "message",
          "code",
          "status"
        ]
      },
...
      "CreateFavoriteRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the media item to be added to favorites",
            "required": true,
            "example": "Inception"
          },
          "year": {
            "type": "string",
            "description": "Year of release of the media item",
            "required": true,
            "example": "2010"
          },
          "imdbID": {
            "type": "string",
            "description": "IMDb ID of the media item",
            "required": true,
            "example": "tt1375666"
          },
          "type": {
            "type": "string",
            "description": "Type of the media item (e.g., movie, series)",
            "required": true,
            "example": "movie"
          },
          "poster": {
            "type": "string",
            "description": "URL of the poster image for the media item",
            "required": true,
            "example": "https://example.com/poster.jpg"
          }
        },
        "required": [
          "title",
          "year",
          "imdbID",
          "type",
          "poster"
        ]
      },
      "CreateFavoriteResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "60c72b2f9b1e8b001c8e4d3a"
          },
          "title": {
            "type": "string",
            "example": "Inception"
          },
          "year": {
            "type": "string",
            "example": "2010"
          },
          "imdbID": {
            "type": "string",
            "example": "tt1375666"
          },
          "type": {
            "type": "string",
            "example": "movie"
          },
          "poster": {
            "type": "string",
            "example": "https://example.com/poster.jpg"
          }
        }
      },
      "GetFavoritesResponse": {
        "type": "object",
        "properties": {
          "favorites": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "60c72b2f9b1e8b001c8e4d3a"
                },
                "title": {
                  "type": "string",
                  "example": "Inception"
                },
                "year": {
                  "type": "string",
                  "example": "2010"
                },
                "imdbID": {
                  "type": "string",
                  "example": "tt1375666"
                },
                "type": {
                  "type": "string",
                  "example": "movie"
                },
                "poster": {
                  "type": "string",
                  "example": "https://example.com/poster.jpg"
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time",
                  "example": "2021-06-14T10:00:00Z"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time",
                  "example": "2021-06-14T10:00:00Z"
                },
                "__v": {
                  "type": "integer",
                  "example": 0
                },
                "description": {
                  "type": "string",
                  "example": "A mind-bending thriller about dreams within dreams."
                }
              }
            }
          },
          "totalFavorites": {
            "type": "integer",
            "example": 1
          },
          "currentPage": {
            "type": "integer",
            "example": 1
          },
          "pageSize": {
            "type": "integer",
            "example": 10
          }
        }
      },
      "DeleteFavoriteResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Favorite deleted successfully"
          },
          "data": {
            "type": "object",
            "description": "Empty object indicating successful deletion",
            "example": {}
          }
        }
      },
      "UpdateFavoriteRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          }
        },
        "required": [
          "description"
        ]
      },
...
      "UserNotFound": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "User not found"
          },
          "code": {
            "type": "string",
            "example": "USER_NOT_FOUND"
          },
          "status": {
            "type": "integer",
            "example": 404
          }
        },
        "required": [
          "error",
          "code",
          "status"
        ]
      },
      "FavoriteNotFound": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Favorite not found"
          },
          "code": {
            "type": "string",
            "example": "FAVORITE_NOT_FOUND"
          },
          "status": {
            "type": "integer",
            "example": 404
          }
        },
        "required": [
          "error",
          "code",
          "status"
        ]
      },
      "FavoriteAlreadyExists": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Favorite already exist for this user"
          },
          "code": {
            "type": "string",
            "example": "FAVORITE_EXISTS"
          },
          "status": {
            "type": "integer",
            "example": 409
          }
        }
      }
    },
    "parameters": {
...
      "UserIdParam": {
        "name": "userId",
        "in": "path",
        "required": true,
        "example": "60c72b2f9b1e8b001c8e4d3a",
        "schema": {
          "type": "string"
        },
        "description": "ID of the user to retrieve favorites for."
      },
      "MediaItemIdParams": {
        "name": "mediaItemId",
        "in": "path",
        "required": true,
        "example": "6847f9acf8a5c5ba5138bf2b",
        "schema": {
          "type": "string"
        },
        "description": "ID of the media item to add or update in favorites."
      },
      "FavoritesPageParam": {
        "name": "page",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1
        },
        "description": "Page number for pagination of favorites (default: 1)."
      },
      "FavoritesPageSizeParam": {
        "name": "pageSize",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1
        },
        "description": "Number of favorites to return per page (default: 10)."
      },
      "FavoritesSortFieldParam": {
        "name": "sort",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "enum": [
            "title",
            "year"
          ]
        },
        "description": "Field to sort favorites by (default: title)."
      },
      "FavoritesSortOrderParam": {
        "name": "order",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "enum": [
            1,
            -1
          ],
          "default": 1
        },
        "description": "Sort order (1 for ascending, -1 for descending; default: 1)."
      },
      "FavoritesMediaTypeParam": {
        "name": "mediaType",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "enum": [
            "movie",
            "series",
            "game",
            "all"
          ],
          "default": "all"
        },
        "description": "Filter favorites by media type (movie, series, game, or all; default: all)."
      },
      "FavoritesSearchTermParam": {
        "name": "searchTerm",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Search term to filter favorites by title."
      }
    }
  },
  "paths": {
    "/api/favorites/{userId}": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Favorites"
        ],
        "summary": "Create a new favorite",
        "description": "Create a new favorite for a user.",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFavoriteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Favorite created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateFavoriteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                },
                "example": {
                  "error": "Validation failed",
                  "code": "VALIDATION_ERROR",
                  "status": 400,
                  "errors": [
                    {
                      "msg": "title is required",
                      "param": "title",
                      "location": "body"
                    },
                    {
                      "msg": "imdbID must be a valid string",
                      "param": "imdbID",
                      "location": "body"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotFound"
                }
              }
            }
          },
          "409": {
            "description": "Favorite already exists for this user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteAlreadyExists"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      },
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Favorites"
        ],
        "summary": "Get user's favorites",
        "description": "Retrieve a paginated list of a user's favorites, total count, current page, and page size with optional filtering and sorting.",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdParam"
          },
          {
            "$ref": "#/components/parameters/FavoritesPageParam"
          },
          {
            "$ref": "#/components/parameters/FavoritesPageSizeParam"
          },
          {
            "$ref": "#/components/parameters/FavoritesSortFieldParam"
          },
          {
            "$ref": "#/components/parameters/FavoritesSortOrderParam"
          },
          {
            "$ref": "#/components/parameters/FavoritesMediaTypeParam"
          },
          {
            "$ref": "#/components/parameters/FavoritesSearchTermParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Favorites retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetFavoritesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                },
                "example": {
                  "error": "Validation failed",
                  "code": "VALIDATION_ERROR",
                  "status": 400,
                  "errors": [
                    {
                      "msg": "userId must be a valid MongoDB ObjectId",
                      "param": "userId",
                      "location": "path"
                    },
                    {
                      "msg": "page must be a positive integer",
                      "param": "page",
                      "location": "query"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotFound"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/api/favorites/{id}/{userId}": {
      "delete": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Favorites"
        ],
        "summary": "Delete a favorite",
        "description": "Deletes a favorite item by its ID for a specific user. If the favorite is the last one associated with the media item, the media item is also deleted.",
        "parameters": [
          {
            "$ref": "#/components/parameters/MediaItemIdParams"
          },
          {
            "$ref": "#/components/parameters/UserIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Favorite deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteFavoriteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                },
                "example": {
                  "error": "Validation failed",
                  "code": "VALIDATION_ERROR",
                  "status": 400,
                  "errors": [
                    {
                      "msg": "id must be a valid MongoDB ObjectId",
                      "param": "id",
                      "location": "path"
                    },
                    {
                      "msg": "userId must be a valid MongoDB ObjectId",
                      "param": "userId",
                      "location": "path"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Favorite or user not found",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/FavoriteNotFound"
                    },
                    {
                      "$ref": "#/components/schemas/UserNotFound"
                    }
                  ]
                },
                "examples": {
                  "FavoriteNotFound": {
                    "summary": "Favorite not found",
                    "value": {
                      "status": 404,
                      "message": "Favorite not found",
                      "code": "FAVORITE_NOT_FOUND"
                    }
                  },
                  "UserNotFound": {
                    "summary": "User not found",
                    "value": {
                      "status": 404,
                      "message": "User not found",
                      "code": "USER_NOT_FOUND"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Favorites"
        ],
        "summary": "Update a favorite",
        "description": "\nUpdates the description of a favorite item for a specific user.\n- If the favorite has no description, this endpoint will add one.\n- If the request body contains an empty string (\"\"), the existing description will be removed.\n- If the request body contains a new string, it will overwrite the previous description.\nThe description is a user-generated note about the favorite media item.\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/MediaItemIdParams"
          },
          {
            "$ref": "#/components/parameters/UserIdParam"
          }
        ],
        "requestBody": {
          "description": "Description of the favorite item to be updated. Can be an empty string to remove the description.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFavoriteRequest"
              },
              "examples": {
                "addOrUpdate": {
                  "summary": "Add or update description",
                  "value": {
                    "description": "A mind-bending thriller about dreams within dreams."
                  }
                },
                "remove": {
                  "summary": "Remove description",
                  "value": {
                    "description": ""
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Favorite updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateFavoriteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                },
                "example": {
                  "error": "Validation failed",
                  "code": "VALIDATION_ERROR",
                  "status": 400,
                  "errors": [
                    {
                      "msg": "id must be a valid MongoDB ObjectId",
                      "param": "id",
                      "location": "path"
                    },
                    {
                      "msg": "userId must be a valid MongoDB ObjectId",
                      "param": "userId",
                      "location": "path"
                    },
                    {
                      "msg": "Description must be a string",
                      "param": "description",
                      "location": "body"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Favorite or user not found",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/FavoriteNotFound"
                    },
                    {
                      "$ref": "#/components/schemas/UserNotFound"
                    }
                  ]
                },
                "examples": {
                  "FavoriteNotFound": {
                    "summary": "Favorite not found",
                    "value": {
                      "error": "Favorite not found or could not be updated",
                      "code": "FAVORITE_NOT_FOUND",
                      "status": 404
                    }
                  },
                  "UserNotFound": {
                    "summary": "User not found",
                    "value": {
                      "error": "User not found",
                      "code": "USER_NOT_FOUND",
                      "status": 404
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}